Update broken CPERF Dockerfile
seems newer CentOS versions (e.g. 7.5) need to have their
distro "centos" available in a yum vars file when trying
to work with SIG repositories

also updated the openstack repo to queens release from
the older ocata

finally, updated the repo location for openflow and oflops.
the stanford repo no longer works, but the github repos
claim to be mirrors of it.

Change-Id: I7084c38c680e1875558af17a3ef3c2344ee9cefc
Signed-off-by: Jamo Luhrsen <jluhrsen@redhat.com>