Update patch set 1
Patch Set 1:

> Build Failed
 > 
 > https://build.opnfv.org/ci/job/opnfv-security-audit-verify-master/17156/
 > : FAILURE (skipped)

This is failing because the security auditing tool blacklists things like git clones from non-OPNFV repos. We talked about this with Tim last time and the advice was to just ignore it. If we really wanted to do the right thing there's apparently a way to add it to the security tool's whitelist, but I don't know how and I don't think it's worth the effort at the moment.

Patch-set: 1